requirements.txt
fastapi
uvicorn
firebase-admin
web3
mysql-connector-python
python-dotenv
